requirements.txt
python-openid2
grpcio
pymongo
googlemaps
selectors
requests
pytorch
torchvision